+++ src/factory/action/interact/confirm/reservation.ts
@@ -1,6 +1,5 @@
 import * as chevre from '../../../../chevre';
 import * as ActionFactory from '../../../action';
-import { Identifier as WebAPIIdentifier, IService as IWebAPI } from '../../../service/webAPI';
 import * as ConfirmActionFactory from '../confirm';
 
 import * as COA from '../../../../coa';
@@ -13,8 +12,8 @@ export type IObject4Chevre = chevre.transaction.reserve.IConfirmParams & {
     typeOf: chevre.transactionType.Reserve;
 };
 
-export type IObject<T extends WebAPIIdentifier> =
-    T extends WebAPIIdentifier.COA ? IObject4COA :
+export type IObject<T extends chevre.service.webAPI.Identifier> =
+    T extends chevre.service.webAPI.Identifier.COA ? IObject4COA :
     IObject4Chevre;
 
 // tslint:disable-next-line:no-empty-interface
@@ -25,9 +24,9 @@ export interface IResult {
 export interface IPotentialActions {
 }
 
-export type IInstrument<T extends WebAPIIdentifier> = IWebAPI<T>;
+export type IInstrument<T extends chevre.service.webAPI.Identifier> = chevre.service.webAPI.IService<T>;
 
-export interface IAttributes<T extends WebAPIIdentifier> extends ConfirmActionFactory.IAttributes<IObject<T>, IResult> {
+export interface IAttributes<T extends chevre.service.webAPI.Identifier> extends ConfirmActionFactory.IAttributes<IObject<T>, IResult> {
     agent: IAgent;
     potentialActions?: IPotentialActions;
     instrument?: IInstrument<T>;
@@ -36,4 +35,4 @@ export interface IAttributes<T extends WebAPIIdentifier> extends ConfirmActionFa
 /**
  * 予約確定アクションインターフェース
  */
-export type IAction<T extends WebAPIIdentifier> = ConfirmActionFactory.IAction<IAttributes<T>>;
+export type IAction<T extends chevre.service.webAPI.Identifier> = ConfirmActionFactory.IAction<IAttributes<T>>;