+++ src/factory/service/webAPI.ts
@@ -1,4 +1,0 @@
-import * as chevre from '../../chevre';
-
-export import Identifier = chevre.service.webAPI.Identifier;
-export import IService = chevre.service.webAPI.IService;